Bledevice and commandpage modified added status_page

+++ sofia_test_app/lib/models/BLESample.dart
@@ -2,14 +2,13 @@ import 'dart:core';
 
 import 'package:sofia_test_app/enums/ble_device_type.dart';
 
-
 class BLESample {
   String deviceId;
   BleDeviceType deviceType;
   String alias;
   DateTime timestamp;
-  int? txPower;
-  int? rxPower;
+  double? txPower;
+  double? rxPower;
 
   BLESample({
     required this.deviceId,
@@ -20,7 +19,6 @@ class BLESample {
     this.rxPower,
   });
 
-  
   @override
   String toString() {
     return 'BLESample{'
